refactor(userportal): move admin buttons UI to pages/admin_buttons.rs, dedupe button form parsing (S1-J)
Scoped slice of the userportal god-module finding: admin_buttons_page,
render_buttons_table, the button CRUD handlers, the SEC-058 is_safe_dom_id
guard, and ICON_OPTIONS move from mod.rs into pages/admin_buttons.rs,
matching the pages/ convention the other multi-page blocks follow. The
branding-settings page intentionally stays in mod.rs — S2-M absorbs it
into the shared settings-form component.

Also extracts parse_button_form (the byte-identical 30-line parse/validate/
json_map block shared by handle_create_button and handle_update_button) and
buttons_table_response (the htmx table re-render tail shared by all three
mutating handlers). mod.rs: 860 -> 527 lines.
