+++ crates/solobase-core/src/blocks/userportal/pages/mod.rs
@@ -1,3 +1,4 @@
+pub mod admin_buttons;
 pub mod dashboard;
 pub mod profile;
 pub mod security;
